requirements.txt
Cython==0.29.6
dwave-ocean-sdk==3.1.1
h5py==2.9.0
matplotlib==3.3.2
networkx==2.5
nltk==3.4.5
nose==1.3.7
numpy==1.16.2
pandas==0.25.0
scikit-learn==0.20.3
scikit-optimize==0.5.2
scipy==1.2.1
seaborn==0.11.0
tqdm==4.31.1